final-pjt-back/requirements.txt
asgiref==3.3.1
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2020.4
sqlparse==0.4.1
psycopg2==2.8.6
django-extensions==3.0.9
requests==2.25.0